requirements.txt
aiohttp==3.9.5
aiomysql==0.2.0
aiosignal==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
allure-pytest==2.9.45
allure-python-commons==2.9.45
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.5.2
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==23.2.0
awscli-local==0.22.2
backports.zoneinfo==0.2.1
billiard==4.2.4
bitarray==2.9.2
blinker==1.7.0
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.2
cast==0.1.1
casttube==0.2.1
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
ChatService==1.0.0
ci-info==0.3.0
ckzg==2.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.4
configobj==5.0.9
configparser==7.0.0
crcmod==1.7
cryptography==41.0.7
cytoolz==0.12.3
decorator==5.1.1
deepdiff==4.3.2
DingtalkChatbot==1.5.7
distlib==0.3.6
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
django-sslserver==0.22
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docker==7.0.0
docopt==0.6.2
durationpy==0.10
elastic-transport==8.17.1
elasticsearch==8.19.3
ens==0.6.1
et-xmlfile==1.1.0
etelemetry==0.3.1
eth-account==0.13.7
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.1.0
eth-tester==0.1.0b11
eth-typing==5.0.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.2.2
fastapi==0.124.4
filelock==3.8.0
fitz==0.0.1.dev2
Flask==3.0.2
foundry==0.0.3
frozenlist==1.4.1
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
gspread==6.2.1
h11==0.16.0
hexbytes==1.2.1
HTMLTestRunner-Python3==0.8.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.7
ifaddr==0.2.0
importlib-metadata==4.11.3
importlib_resources==6.4.0
iniconfig==2.0.0
install==1.3.5
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.9.1
jmespath==0.10.0
jsonpath==0.82
jsonpath-rw==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.5.4
kubernetes==34.1.0
localstack-client==2.10
looseversion==1.3.0
lru-dict==1.2.0
lxml==4.9.1
MarkupSafe==2.1.5
meteo-downloader==1.0
motor==3.3.2
multidict==6.0.5
mysql==0.0.3
mysql-connector-python==8.2.0
mysql-connector-python-rf==2.2.2
mysql-replication==1.0.12
mysqlclient==2.1.0
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==2.2.0
openpyxl==3.0.9
ordered-set==4.0.2
oss2==2.18.4
outcome==1.3.0.post0
packaging==24.1
pandas==2.0.3
parsimonious==0.10.0
pathlib==1.0.1
pika==1.3.2
pika-stubs==0.1.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.5.0
ply==3.11
prompt_toolkit==3.0.52
protobuf==4.21.12
prov==2.0.1
psutil==7.1.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyChromecast==13.1.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.9.2
pydantic-settings==2.8.1
pydantic_core==2.23.4
pydot==4.0.1
PyJWT==1.7.1
pymongo==4.6.1
PyMySQL==1.1.0
pyparsing==3.1.4
pysha3==1.0.2
PySocks==1.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pytz-deprecation-shim==0.1.0.post0
pyunormalize==15.1.0
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==6.3.2
redis==5.0.3
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.1.0
rpds-py==0.19.1
rsa==4.9.1
s3transfer==0.11.5
scipy==1.10.1
selenium==4.22.0
semantic-version==2.10.0
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.2
starlette==0.44.0
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.1
traits==6.3.2
trio==0.26.0
trio-websocket==0.11.1
tweepy==4.14.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.33.0
vine==5.1.0
virtualenv==20.16.5
vlog==2.1.0
wcwidth==0.2.14
web3==7.13.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
xlrd==1.2.0
Yamler==0.0.1
yarl==1.9.4
zeroconf==0.132.2
zipp==3.8.0
gunicorn==21.2.0
